Cart functions done

+++ ecommerce-backend/src/main/java/com/saravanank/ecommerce/resourceserver/repository/CartRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.saravanank.ecommerce.resourceserver.model.Cart;
 
-public interface CartRepository extends JpaRepository<Cart, Integer> {
+public interface CartRepository extends JpaRepository<Cart, Long> {
 
 }
